requirements.txt
pefile==2023.2.7
python-magic==0.4.18
python-magic-bin==0.4.18